requirements.txt
# Core Django (Latest LTS)
Django==5.0.8
djangorestframework==3.15.2
django-cors-headers==4.4.0

# Database (Modern PostgreSQL adapter)
psycopg[binary]==3.2.1
django-extensions==3.2.3

# Authentication & Security
djangorestframework-simplejwt==5.3.0
django-allauth==0.63.6
cryptography==42.0.8

# Environment Management
python-dotenv==1.0.1
django-environ==0.11.2

# API Documentation
drf-spectacular==0.27.2

# File Storage & Media
django-storages==1.14.4
Pillow==10.4.0

# Caching & Session
redis==5.0.8
django-redis==5.4.0

# Task Queue
celery==5.4.0

# Real-time Features
channels==4.1.0
channels-redis==4.2.0

# Development Tools
django-debug-toolbar==4.4.6
ipython==8.26.0

# Testing
pytest==8.3.2
pytest-django==4.8.0
factory-boy==3.3.0
pytest-cov==5.0.0

# Code Quality
black==24.8.0
ruff==0.5.7
pre-commit==3.8.0